pyproject.toml
[tool.poetry]
name = "dcw"
version = "0.0.7"
description = ""
authors = ["Derek Wisong <derekwisong@gmail.com>"]
readme = "README.md"
packages = [
    { include = "dcw", from = "." },
]

[tool.poetry.scripts]
dcw-pipeline = "dcw.cli.pipeline:main"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.1.3"
pytest = "^7.4.3"
requests = "^2.31.0"
pydantic = "^2.5.2"
streamz = "^0.6.4"
google-cloud-storage = "^2.14.0"
tabulate = "^0.9.0"
python-snappy = "^0.6.1"
pyarrow = "^14.0.2"
fastparquet = "^2023.10.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
autopep8 = "^2.0.4"
mkdocs = "^1.5.3"
mkdocs-material = "^9.5.2"
mkdocstrings = {extras = ["python"], version = "^0.24.0"}
mkdocs-gen-files = "^0.5.0"
mkdocs-literate-nav = "^0.6.1"
mkdocs-section-index = "^0.3.8"
markdown-include = "^0.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
